requirements.txt
altgraph==0.17.2
future==0.18.2
importlib-metadata==4.11.0
pefile==2021.9.3
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.1
pywin32-ctypes==0.2.0
typing_extensions==4.1.1
zipp==3.7.0

pyzmail36
IMAPClient~=2.3.1
customtkinter~=5.1.2